pyproject.toml
[project]
name = "news_extraction"
version = "0.1.0"
description = "News scraping and summarization pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.51.0",
    "openai>=1.68.2",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.1",
    "lxml[html_clean]>=5.3.1",
    "beautifulsoup4>=4.12.3",
    "loguru>=0.7.2",
    "uvicorn>=0.29.0",
    "feedparser>=6.0.12",
    "newspaper3k>=0.2.8",
    "python-telegram-bot>=22.4",
    "scikit-learn>=1.7.2",
    "pyttsx3>=2.99",
    "pydub>=0.25.1",
    "mcp>=1.14.1",
    "pandas>=2.3.2",
    "flask>=3.0.0",
    "google-cloud-storage>=2.20.0",
    "google-cloud-texttospeech>=2.30.0",
    "gunicorn>=23.0.0",
]

[build-system]
requires = ["setuptools>=78.1.0"]
build-backend = "setuptools.build_meta"
